+++ save-orchestrator-common/src/main/kotlin/com/saveourtool/save/orchestrator/controller/AgentsController.kt
@@ -4,7 +4,6 @@ import com.saveourtool.save.execution.ExecutionStatus
 import com.saveourtool.save.orchestrator.runner.ContainerRunnerException
 import com.saveourtool.save.orchestrator.service.AgentService
 import com.saveourtool.save.orchestrator.service.ContainerService
-import com.saveourtool.save.orchestrator.service.ExecutionService
 import com.saveourtool.save.request.RunExecutionRequest
 import com.saveourtool.save.utils.EmptyResponse
 import com.saveourtool.save.utils.info
@@ -25,7 +24,6 @@ import reactor.core.publisher.Mono
 @RestController
 class AgentsController(
     private val agentService: AgentService,
-    private val executionService: ExecutionService,
     private val containerService: ContainerService,
 ) {
     /**
@@ -38,7 +36,7 @@ class AgentsController(
     @Suppress("TOO_LONG_FUNCTION", "LongMethod", "UnsafeCallOnNullableType")
     @PostMapping("/initializeAgents")
     fun initialize(@RequestBody request: RunExecutionRequest): Mono<EmptyResponse> {
-        val response = executionService.updateExecution(request.executionId, ExecutionStatus.INITIALIZATION)
+        val response = agentService.updateExecution(request.executionId, ExecutionStatus.INITIALIZATION)
             .thenReturn(ResponseEntity<Void>(HttpStatus.ACCEPTED))
             .subscribeOn(agentService.scheduler)
         return response.doOnSuccess {
@@ -57,7 +55,7 @@ class AgentsController(
                 }
                 .flatMap {
                     log.info("Sending request to make execution.id=${request.executionId} RUNNING")
-                    executionService.updateExecution(request.executionId, ExecutionStatus.RUNNING)
+                    agentService.updateExecution(request.executionId, ExecutionStatus.RUNNING)
                 }
                 .flatMap {
                     containerService.validateContainersAreStarted(request.executionId)
@@ -71,7 +69,7 @@ class AgentsController(
         ex: ContainerRunnerException,
     ): Mono<T> {
         log.error("${ex.message} for executionId=$executionId, will mark it as ERROR", ex)
-        return executionService.updateExecution(executionId, ExecutionStatus.ERROR, ex.message)
+        return agentService.updateExecution(executionId, ExecutionStatus.ERROR, ex.message)
             .then(Mono.empty())
     }
